Relocate CreateStroll.jsx to components
